backend/prediction-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dateutil==2.8.2
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
httpx==0.25.2
python-multipart==0.0.6

